split tests for notifications
